+++ addons/eu/en16931/en16931.go
@@ -39,6 +39,7 @@ func init() {
 		orgAttachmentRules(),
 		orgPartyRules(),
 		orgInboxRules(),
+		orgEndpointRules(),
 		orgAddressRules(),
 		taxComboRules(),
 	)